Add byte type (TypeCode=17): 0-255 with wrapping arithmetic
- Value::Byte(u8), TypeCode::Byte=17
- byte + byte = byte (mod 256 wrapping): 255+1=0, 0-1=255
- byte + int = int (byte promotes); byte ^ byte = byte
- byte == int cross-type equality (byte(65)==65 is true)
- byteArray/bytes index returns Byte (not Int)
- byteArray accepts Byte or Int(0-255) for index set
- ~byte = byte (8-bit not)
- New builtins: byte(), isByte(), bytesXor(data,key), bytesXorInPlace(ba,key)
- bytesXor: efficient Rust-level XOR loop for encryption (vs slow script loop)
- Updated existing tests for Byte return type
- 117 integration + 25 unit tests, all passing

+++ sflang/src/builtins_str.rs
@@ -238,7 +238,7 @@ fn bi_bytes_at(_vm: &mut VM, args: &[Value]) -> Result<Value, Value> {
             "bytesAt() 索引 {} 越界 (len={}); 可能原因：索引超出字节数", i, n,
         )));
     }
-    Ok(Value::Int(bytes[i as usize] as i64))
+    Ok(Value::Byte(bytes[i as usize]))
 }
 
 /// bi_len_bytes 返回 string 的 UTF-8 字节数。
